requirements.txt
BitTorrent==3.4.2
Brlapi==0.5.6
CouchDB==0.8
Flask==0.9
Flask-Assets==0.8
GnuPGInterface==0.3.2
Jinja2==2.7
Mako==0.5.0
MarkupSafe==0.18
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
PyDispatcher==2.0.3
PyOpenGL==3.0.1b2
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
Werkzeug==0.9.1
adium-theme-ubuntu==0.3.2
apparmor==2.7.102
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
blinker==1.3
ccsm==0.9.5.92
chardet==2.0.1
command-not-found==0.2.44
compizconfig-python==0.9.5.94
configglue==1.0
configobj==4.7.2
debtagshw==0.1
decorator==3.3.2
defer==1.0.2
dirspec==3.0.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
dnspython==1.9.4
doge==3.4.0
duplicity==0.6.18
feedparser==5.1
httplib2==0.7.2
ipython==0.12.1
itsdangerous==0.22
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
logilab-astng==0.23.0
logilab-common==0.57.1
louis==2.3.0
lxml==3.2.1
mercurial==2.0.2
numpy==1.6.1
nvidia-common==0.0.0
oauth==1.0.1
onboard==0.97.1
oneconf==0.2.8.1
openshot==1.4.0
pexpect==2.3
piston-mini-client==0.7.2
pithos==0.3
protobuf==2.4.1
pyOpenSSL==0.12
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pygame==1.9.1release
pyinotify==0.9.2
pylint==0.25.0
pyogg==1.3
pyserial==2.5
pysmbc==1.0.13
python-apt==0.8.3ubuntu7.1
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-virtkey==0.60.0
pyvorbis==1.5
pyxdg==0.19
pyzmq==2.1.11
rabbitvcs==0.13.1
rdflib==2.4.2
readline==6.2.4.1
reportlab==2.5
rhythmbox-ubuntuone==4.2.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
simplegeneric==0.7
simplejson==2.3.2
software-center-aptd-plugins==0.0.0
stevedore==0.12
system-service==0.1.6
tornado==2.1
uTidylib==0.2
uWSGI==1.9.18.2
ubuntu-tweak==0.8.6
ubuntuone-couch==0.3.0
ubuntuone-installer==3.0.2
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
urlgrabber==3.9.1
usb-creator==0.2.23
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.0
webassets==0.8
wsgi-intercept==0.4
wsgiref==0.1.2
wxPython==2.6.3.2
wxPython-common==2.6.3.2
xdiagnose==2.5.3
xkit==0.0.0
xmlrpclib==1.0.1
yolk==0.4.3
zenmap==5.21
zim==0.54
zope.interface==3.6.1
